Atualização da padronização dos métodos

+++ src/main/java/br/com/betuka/automec/service/tabApoio/MarcaService.java
@@ -2,6 +2,7 @@ package br.com.betuka.automec.service.tabApoio;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Objects;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -22,28 +23,53 @@ public class MarcaService {
 		return MarcaDTO.toList(this.marcaRepository.findAll());
 	}
 	
-	public void validarEntrada(MarcaDTO pMarcaDTO) throws ValidationException, Exception {
-		if (pMarcaDTO.getCodMarca() != 0) {
-			this.pesquisarCodigo(pMarcaDTO.getCodMarca()); // Caso não encontre já levanta uma Exception
+	public void validarEntrada(MarcaDTO marcaDTO) throws ValidationException, Exception {
+		if (marcaDTO.getCodMarca() != 0) {
+			this.pesquisarCodigo(marcaDTO.getCodMarca()); // Caso não encontre já levanta uma Exception
 		}
 		
-		if (pMarcaDTO.getDesMarca().isEmpty()) {
+		if (marcaDTO.getDesMarca().isEmpty()) {
 			throw new ValidationException(Constants.MARCA_DESCRICAO_OBRIGATORIA);
+		}
+		
+		MarcaDTO oMarcaDTO = null;
+		
+		try {
+			oMarcaDTO = this.pesquisarDescricao(marcaDTO.getDesMarca());
+		} catch (ValidationException e) {
+			// Neste caso não trata a exceção
+		}
+		
+		if (Objects.nonNull(oMarcaDTO)) {
+			if (marcaDTO.getCodMarca() == 0) {
+				throw new ValidationException(Constants.MARCA_JA_CADASTRADO);
+			} else {
+				if (marcaDTO.getCodMarca() != oMarcaDTO.getCodMarca()) {
+					throw new ValidationException(Constants.MARCA_JA_CADASTRADO);
+				}
+			}
 		}
 	}
 	
-	public void adicionar(MarcaDTO pMarcaDTO) throws ValidationException, Exception {
-		this.validarEntrada(pMarcaDTO);
-		this.marcaRepository.save( new MarcaEntity(pMarcaDTO) );
+	public void gravar(MarcaDTO marcaDTO) throws ValidationException, Exception {
+		this.validarEntrada(marcaDTO);
+		this.marcaRepository.save( new MarcaEntity(marcaDTO) );
 	}
-	
-	public void atualizar(MarcaDTO pMarcaDTO) throws ValidationException, Exception {
-		this.validarEntrada(pMarcaDTO);
-		this.marcaRepository.save( new MarcaEntity(pMarcaDTO) );
-	}
+    
+    private boolean existeMarca(int codMarca) throws ValidationException, Exception {
+    	try {
+    		return this.marcaRepository.existeMarca(codMarca);
+		} catch (Exception e) {
+			throw new Exception(e.getMessage());
+		}
+    }
 	
 	public void deletar(int codMarca) throws ValidationException, Exception {
 		this.pesquisarCodigo(codMarca); // Caso não encontre levanta uma ValidationException
+		
+		if (this.existeMarca(codMarca)) {
+			throw new ValidationException(Constants.MARCA_UTILIZADA);
+		}
 		
 		try {
 			this.marcaRepository.deleteById(codMarca);	
@@ -52,9 +78,9 @@ public class MarcaService {
 		}
 	}
 	
-	public MarcaDTO pesquisarCodigo(int pCodMarca) throws ValidationException, Exception {
+	public MarcaDTO pesquisarCodigo(int codMarca) throws ValidationException, Exception {
 		try {
-			return new MarcaDTO(this.marcaRepository.findById(pCodMarca).get());
+			return new MarcaDTO(this.marcaRepository.findById(codMarca).get());
 		} catch (NoSuchElementException e) {
 			throw new ValidationException(Constants.MARCA_INEXISTENTE);
 		} catch (Exception e) {
@@ -62,11 +88,19 @@ public class MarcaService {
 		}
 	}
 	
-	public MarcaDTO pesquisarDescricao(String pDesMarca) throws ValidationException, Exception {
+	public MarcaDTO pesquisarDescricao(String desMarca) throws ValidationException, Exception {
 		try {
-			return new MarcaDTO(this.marcaRepository.pesquisarDescricao(pDesMarca).get());
+			return new MarcaDTO(this.marcaRepository.pesquisarDescricao(desMarca).get());
 		} catch (NoSuchElementException e) {
 			throw new ValidationException(Constants.MARCA_INEXISTENTE);
+		} catch (Exception e) {
+			throw new Exception(e.getMessage());
+		}
+	}
+	
+	public List<MarcaDTO> pesquisarPorDescricao(String desMarca) throws ValidationException, Exception {
+		try {
+			return MarcaDTO.toList(this.marcaRepository.pesquisarPorDescricao(desMarca));
 		} catch (Exception e) {
 			throw new Exception(e.getMessage());
 		}
